mnist_project/requirements.txt
# local package
-e .

numpy
matplotlib
scikit-learn
torch
torchvision
torchaudio

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
